refactor user queries
