+++ src/api/users/__tests__/userpage.test.ts
@@ -1,5 +1,5 @@
 import endpoint from '../../../../pages/api/users/[id]';
-import { users } from '../../../../db/queries/users';
+import { getUserData } from '../../../../db/queries/users';
 
 jest.mock('../../../../db/queries/users');
 
@@ -24,14 +24,14 @@ describe('users /[id]', () => {
   };
 
   it('return success if db call is successful', async () => {
-    users.getUserData.mockImplementationOnce((id) => Promise.resolve());
+    getUserData.mockImplementationOnce((id) => Promise.resolve());
 
     const response = await endpoint(mockReq, mockRes());
     expect(response.status).toEqual(200);
   });
 
   it('return failure if db call is unsuccessful', async () => {
-    users.getUserData.mockImplementationOnce((id) => Promise.reject());
+    getUserData.mockImplementationOnce((id) => Promise.reject());
 
     const response = await endpoint(mockReq, mockRes());
     expect(response.status).toEqual(404);
