+++ src/helpers/userProfileFetcher.ts
@@ -3,7 +3,7 @@ import userDataFormatter from './userDataFormatter';
 import { getAuth0User } from './auth0/auth0User';
 import patreonProfileFetcher from './patreon/profileFetcher';
 import getAuth0UserSub from './auth0/auth0Sub';
-import userQueries from '../../db/queries/users';
+import { updateUser, getUserData } from '../../db/queries/users';
 import { DisplayUser, PatreonTokenData } from '../types/common';
 import { AvatarSelect } from '../types/enums';
 import { MailchimpSubscriberStatus } from '../types/api/apiTypes.d';
@@ -99,14 +99,14 @@ export default async function userProfileFetcher(req) {
       reads,
       wishlist,
       ratings,
-    } = await userQueries.users.getUserData(userData.onbc_id);
+    } = await getUserData(userData.onbc_id);
 
     //check for mismatch in avatars and correct in db
     if (
       gravatar_avatar_url !== userData.avatar ||
       (patreon_avatar_url !== patreonAvatar && patreonSuccess === true)
     ) {
-      await userQueries.users.update(userData.onbc_id, {
+      await updateUser(userData.onbc_id, {
         gravatar_avatar_url: userData.avatar,
         patreon_avatar_url: patreonAvatar,
       });
